renamed server.js in test to index.js to reflect src/ file structure
